fix: :poop: change type of getter for 1d index to uint16

+++ firmware/Team01.cydsn/lib/Point/Point.c
@@ -58,7 +58,7 @@ uint8_t Point_get_y(Point *point)
 {
 	return point->y;
 }
-uint8_t Point_get_1d(Point *point)
+uint16_t Point_get_1d(Point *point)
 {
 	return point->one_dim_index;
 }
